% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riksbank.R
\name{riksbank_series}
\alias{riksbank_series}
\title{List all available Riksbank SWEA series}
\usage{
riksbank_series()
}
\value{
tibble: seriesid (character), name (character), description (character)
}
\description{
List all available Riksbank SWEA series
}
